Fixing collision bug
This fixes the bug where the player could move through a dynamic object
without triggering a collision if the player and the object's positions
exactly swapped in a turn. This was easiest to do my making
dynamicObject.js be responsible for checking for a collision after the
player has moved but before the object itself has.

Note that it's still a good idea to move the special logic associated
with transport objects from player.js into dynamicObject.js. This would
allow us to get rid of a whole large block of code there.