Starting work on final boss fight
Right now, we're thinking that the final boss fight will be based around
the concept of the code changing as a certain amount of damage to the
robot boss is done. We'll need a health bar for the robot. Dr. Eval's
challenge is to read the code and figure out the pattern so he can
change the function phone to fire back the appropriate response to that
stage of the robot.

IDEA:
Every time the robot gets a new "code weapon", dr eval can see that in
the code and now has access to it.