Create 18_super_Dr_Eval_Bros.md
Unlike the generic auto-build bridge across gap solutions, you actually have to pick up a tile, and use the phone callback to place it, and repeat this 10 times. (You could just make a run for it, but if gravity pulls you down, you are screwed. :) )

Additionally, instead of a bottomless pit, there are spikes at the bottom to kill the player. :)