fix issues #82 , #309 for level 18
As the above issues describe, this security bug is about javascript not being able to process the gravity function fast enough on button continuous press. So we increase the input button delay, decreasing interval for gravity function call and we widen the gap the player has to "jump" so this javascript issue cannot be exploited to pass the level without writing code as the issues describe. If the changes are implemented in the project, the solutions.md for this level must be updated so the players' solutions match the new wider gap and coordinates.