Prevent player from dying prematurely
Closes #416
If map.getPlayer().killedBy(reason) is called during the startLevel
function, it causes the level to immiedately attempt to restart, causing
an infinite loop. For reasons I don't entirely understand, this can
confuse the editor about which lines are editable and break the game
with "ReferenceError: map is not defined" - but either way it's a bad
idea.