Rewrite credits level to not depend on private APIs
The credits level, as currently coded, depends om the private attribute
`map._display`. Access to the display was mostly broken by #433 and can
lead to security issues so, as a predecessor to a future patch that
blocks access to internal attributes, this rewrites the credits level
to no longer need that access.