Add a DOM bonus level
This was originally intended to be an update to level 19 to resolve #65.
I decided it was better as a bonus level at this stage
in the game's lifecycle.

The level is technically trivial (the solution is just `preferBack =
isAtTopLevel` and then chase the adversary through the entire DOM), but
realizing that is difficult (the structure of the DOM means that
preferBack = true loops indefinitely in the class table and preferBack =
false loops between the class table and the Cicero quote)

My comment in #142 was referring to a more complicated version of this
level where there were multiple sub-line editable sections each of which
was too short to write the words `isAtTopLevel` requiring you to
redefine it via JavaScript string concatenations and square brackets.
But that seems too evil now